Move ShardedLmdb into separate crate
