+++ src/rust/engine/Cargo.toml
@@ -43,6 +43,7 @@ members = [
   "resettable",
   "rule_graph",
   "serverset",
+  "sharded_lmdb",
   "tar_api",
   "testutil",
   "testutil/mock",
@@ -74,6 +75,7 @@ default-members = [
   "resettable",
   "rule_graph",
   "serverset",
+  "sharded_lmdb",
   "tar_api",
   "testutil",
   "testutil/mock",
