+++ src/rust/engine/sharded_lmdb/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "sharded_lmdb"
+version = "0.0.1"
+authors = [ "Pants Build <pantsbuild@gmail.com>" ]
+edition = "2018"
+
+[dependencies]
+fs = { path = "../fs" }
+hashing = { path = "../hashing" }
+lmdb = { git = "https://github.com/pantsbuild/lmdb-rs.git", rev = "06bdfbfc6348f6804127176e561843f214fc17f8" }
+log = "0.4"
+tempfile = "3"
